Decouple PlayerWrapper from main
This it can be instantiated in ui
