+++ examples/player/main.rs
@@ -16,18 +16,10 @@ extern crate webrender;
 extern crate winit;
 
 use gleam::gl;
-use ipc_channel::ipc;
 use servo_media::player::frame::{Frame, FrameRenderer};
-use servo_media::player::{Player, PlayerEvent};
-use servo_media::ServoMedia;
 use std::env;
-use std::fs::File;
-use std::io::BufReader;
-use std::io::Read;
 use std::path::Path;
-use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
-use std::thread::Builder;
 #[cfg(not(target_os = "android"))]
 use ui::HandyDandyRectBuilder;
 use webrender::api::*;
@@ -36,111 +28,8 @@ use webrender::api::*;
 #[path = "ui.rs"]
 mod ui;
 
-struct PlayerWrapper {
-    player: Arc<Mutex<Box<Player>>>,
-    shutdown: Arc<AtomicBool>,
-}
-
-impl PlayerWrapper {
-    pub fn new(path: &Path) -> Self {
-        let servo_media = ServoMedia::get().unwrap();
-        let player = Arc::new(Mutex::new(servo_media.create_player()));
-        let file = File::open(&path).unwrap();
-        let metadata = file.metadata().unwrap();
-        player
-            .lock()
-            .unwrap()
-            .set_input_size(metadata.len())
-            .unwrap();
-        let (sender, receiver) = ipc::channel().unwrap();
-        player.lock().unwrap().register_event_handler(sender);
-        let player_ = player.clone();
-        let player__ = player.clone();
-        let shutdown = Arc::new(AtomicBool::new(false));
-        let shutdown_ = shutdown.clone();
-        let shutdown__ = shutdown.clone();
-        Builder::new()
-            .name("File reader".to_owned())
-            .spawn(move || {
-                let player = &player_;
-                let shutdown = &shutdown_;
-                let mut buf_reader = BufReader::new(file);
-                let mut buffer = [0; 8192];
-                while !shutdown.load(Ordering::Relaxed) {
-                    match buf_reader.read(&mut buffer[..]) {
-                        Ok(0) => {
-                            println!("finished pushing data");
-                            break;
-                        }
-                        Ok(size) => {
-                            if let Err(_) = player
-                                .lock()
-                                .unwrap()
-                                .push_data(Vec::from(&buffer[0..size]))
-                            {
-                                break;
-                            }
-                        }
-                        Err(e) => {
-                            eprintln!("Error: {}", e);
-                            break;
-                        }
-                    }
-                }
-            })
-            .unwrap();
-
-        Builder::new()
-            .name("Player event loop".to_owned())
-            .spawn(move || {
-                let player = &player__;
-                let shutdown = &shutdown__;
-                while let Ok(event) = receiver.recv() {
-                    match event {
-                        PlayerEvent::EndOfStream => {
-                            println!("EOF");
-                            break;
-                        }
-                        PlayerEvent::Error => {
-                            println!("Error");
-                            break;
-                        }
-                        PlayerEvent::MetadataUpdated(ref m) => {
-                            println!("Metadata updated! {:?}", m);
-                        }
-                        PlayerEvent::StateChanged(ref s) => {
-                            println!("Player state changed to {:?}", s);
-                        }
-                        PlayerEvent::FrameUpdated => eprint!("."),
-                        PlayerEvent::PositionChanged(_) => (),
-                        PlayerEvent::SeekData(_) => (),
-                        PlayerEvent::SeekDone(_) => (),
-                        PlayerEvent::NeedData => (),
-                        PlayerEvent::EnoughData => (),
-                    }
-                }
-                player.lock().unwrap().stop().unwrap();
-                shutdown.store(true, Ordering::Relaxed);
-            })
-            .unwrap();
-
-        player.lock().unwrap().play().unwrap();
-
-        PlayerWrapper { player, shutdown }
-    }
-
-    fn shutdown(&self) {
-        self.player.lock().unwrap().stop().unwrap();
-        self.shutdown.store(true, Ordering::Relaxed);
-    }
-
-    fn register_frame_renderer(&self, renderer: Arc<Mutex<FrameRenderer>>) {
-        self.player
-            .lock()
-            .unwrap()
-            .register_frame_renderer(renderer);
-    }
-}
+#[path = "player_wrapper.rs"]
+mod player_wrapper;
 
 struct App {
     frame_queue: Vec<Frame>,
@@ -158,7 +47,12 @@ impl App {
 
 #[cfg(not(target_os = "android"))]
 impl ui::Example for App {
-    fn render(&mut self, api: &RenderApi, builder: &mut DisplayListBuilder, txn: &mut Transaction) {
+    fn push_txn(
+        &mut self,
+        api: &RenderApi,
+        builder: &mut DisplayListBuilder,
+        txn: &mut Transaction,
+    ) {
         let frame = if self.frame_queue.is_empty() {
             if self.current_frame.is_none() {
                 return;
@@ -240,9 +134,6 @@ fn main() {
     };
 
     let path = Path::new(filename);
-    let player_wrapper = PlayerWrapper::new(&path);
     let app = Arc::new(Mutex::new(App::new()));
-    player_wrapper.register_frame_renderer(app.clone());
-    ui::main_wrapper(app, None);
-    player_wrapper.shutdown();
+    ui::main_wrapper(app, &path, None);
 }
